Move ChildDashboardI18nTest to i18n_test.rb to avoid merge conflicts
