+++ test/integration/child_flow_test.rb
@@ -23,19 +23,3 @@ class ChildFlowTest < ActionDispatch::IntegrationTest
     assert_select "a[href='#{edit_preferences_path}']"
   end
 end
-
-class ChildDashboardI18nTest < ActionDispatch::IntegrationTest
-  test "gave_you key is present in all supported locales" do
-    %i[en nl it].each do |locale|
-      assert I18n.exists?("child.dashboard.gave_you", locale: locale),
-             "child.dashboard.gave_you is missing in #{locale}"
-    end
-  end
-
-  test "child dashboard renders a link to preferences" do
-    sign_in_as @child
-    get child_dashboard_path
-    assert_response :success
-    assert_select "a[href='#{edit_preferences_path}']"
-  end
-end
